PLCVS/backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
pydantic==2.5.3
python-multipart==0.0.6

# Audio
PyAudio==0.2.14
soundfile==0.12.1
librosa==0.10.1
numpy==1.26.3

# STT
faster-whisper==1.0.1
ctranslate2==4.0.0

# VAD
silero-vad==4.0  # or torch-based

# NLP
spacy==3.7.2
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.36.2

# LLM Fallback
llama-cpp-python==0.2.44

# Database
aiosqlite==0.19.0

# Config
pyyaml==6.0.1

# Logging
structlog==24.1.0

# Report
reportlab==4.0.9
jinja2==3.1.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Utilities
python-dotenv==1.0.0